+++ pytorch_forecasting/models/_base_model_rename_this.py
@@ -1,0 +1,17 @@
+"""Base classes for pytorch-foercasting models."""
+
+from pytorch_forecasting.models.base import (
+    AutoRegressiveBaseModel,
+    AutoRegressiveBaseModelWithCovariates,
+    BaseModel,
+    BaseModelWithCovariates,
+    Prediction,
+)
+
+__all__ = [
+    "AutoRegressiveBaseModel",
+    "AutoRegressiveBaseModelWithCovariates",
+    "BaseModel",
+    "BaseModelWithCovariates",
+    "Prediction",
+]
