+++ pytorch_forecasting/models/base/__init__.py
@@ -1,0 +1,17 @@
+"""Base classes for pytorch-foercasting models."""
+
+from pytorch_forecasting.models.base._base_model import (
+    AutoRegressiveBaseModel,
+    AutoRegressiveBaseModelWithCovariates,
+    BaseModel,
+    BaseModelWithCovariates,
+    Prediction,
+)
+
+__all__ = [
+    "AutoRegressiveBaseModel",
+    "AutoRegressiveBaseModelWithCovariates",
+    "BaseModel",
+    "BaseModelWithCovariates",
+    "Prediction",
+]
