+++ pytorch_forecasting/models/nbeats/_nbeats.py
@@ -10,7 +10,7 @@ from torch import nn
 from pytorch_forecasting.data import TimeSeriesDataSet
 from pytorch_forecasting.data.encoders import NaNLabelEncoder
 from pytorch_forecasting.metrics import MAE, MAPE, MASE, RMSE, SMAPE, MultiHorizonMetric
-from pytorch_forecasting.models.base_model import BaseModel
+from pytorch_forecasting.models.base import BaseModel
 from pytorch_forecasting.models.nbeats.sub_modules import (
     NBEATSGenericBlock,
     NBEATSSeasonalBlock,
